server/requirements.txt
Flask==2.2.2
Flask_Cors==3.0.10
langchain==0.0.351
pinecone_client==2.2.4
python-dotenv==1.0.0
huggingface-hub==0.19.4
InstructorEmbedding==1.0.1
sentence-transformers==2.2.2
gunicorn==21.2.0
werkzeug==2.2.3